oracle_db_monitor_fastapi/requirements.txt
FastAPI==0.95.1
python-oracledb==1.3.1
uvicorn==0.22.0
python-dotenv==0.19.0
